+++ watch/demo/landsat_demodata.py
@@ -2,30 +2,58 @@ import ubelt as ub
 from os.path import join
 
 
-def grab_landsat_item(scene_id=None):
+def grab_landsat_product(product_id=None):
     """
-    Download and cache all items in a given landsat scene.
+    Download and cache all items for a landsat product.
 
     Args:
-        scene_id (str, default=None):
-            the scene id to download (currently NotImplemented).
-            if unspecified, an arbitrary scene is returned.
+        product_id (str, default=None):
+            The product id to download (currently NotImplemented).
+            If unspecified, an arbitrary scene is returned.
 
     Returns:
-        List[str]: list of files associated with this landsat scene
+        Dict[str, object]:
+            groupings of files associated with this landsat product
 
     Example:
         >>> # xdoctest: +REQUIRES(--network)
         >>> from watch.demo.landsat_demodata import *  # NOQA
-        >>> fpaths = grab_landsat_item()
-        >>> print('fpaths = {}'.format(ub.repr2(fpaths, nl=1)))
+        >>> product = grab_landsat_product()
+        >>> print('product = {}'.format(ub.repr2(product, nl=2)))
+        product = {
+            'bands': [
+                '.../LC08/01/037/029/LC08_L1TP_037029_20130602_20170310_01_T1/LC08_L1TP_037029_20130602_20170310_01_T1_B1.TIF',
+                '.../LC08/01/037/029/LC08_L1TP_037029_20130602_20170310_01_T1/LC08_L1TP_037029_20130602_20170310_01_T1_B2.TIF',
+                '.../LC08/01/037/029/LC08_L1TP_037029_20130602_20170310_01_T1/LC08_L1TP_037029_20130602_20170310_01_T1_B3.TIF',
+                '.../LC08/01/037/029/LC08_L1TP_037029_20130602_20170310_01_T1/LC08_L1TP_037029_20130602_20170310_01_T1_B4.TIF',
+                '.../LC08/01/037/029/LC08_L1TP_037029_20130602_20170310_01_T1/LC08_L1TP_037029_20130602_20170310_01_T1_B5.TIF',
+                '.../LC08/01/037/029/LC08_L1TP_037029_20130602_20170310_01_T1/LC08_L1TP_037029_20130602_20170310_01_T1_B6.TIF',
+                '.../LC08/01/037/029/LC08_L1TP_037029_20130602_20170310_01_T1/LC08_L1TP_037029_20130602_20170310_01_T1_B7.TIF',
+                '.../LC08/01/037/029/LC08_L1TP_037029_20130602_20170310_01_T1/LC08_L1TP_037029_20130602_20170310_01_T1_B8.TIF',
+                '.../LC08/01/037/029/LC08_L1TP_037029_20130602_20170310_01_T1/LC08_L1TP_037029_20130602_20170310_01_T1_B9.TIF',
+                '.../LC08/01/037/029/LC08_L1TP_037029_20130602_20170310_01_T1/LC08_L1TP_037029_20130602_20170310_01_T1_B10.TIF',
+                '.../LC08/01/037/029/LC08_L1TP_037029_20130602_20170310_01_T1/LC08_L1TP_037029_20130602_20170310_01_T1_B11.TIF',
+            ],
+            'meta': {
+                'bqa': '.../LC08/01/037/029/LC08_L1TP_037029_20130602_20170310_01_T1/LC08_L1TP_037029_20130602_20170310_01_T1_BQA.TIF',
+                'mtl': '.../LC08/01/037/029/LC08_L1TP_037029_20130602_20170310_01_T1/LC08_L1TP_037029_20130602_20170310_01_T1_MTL.txt',
+            },
+        }
+
+    References:
+        .. [1] https://github.com/dgketchum/Landsat578#-1
+        .. [2] https://tilemill-project.github.io/tilemill/docs/guides/landsat-8-imagery/
+        .. [3] https://earth.esa.int/documents/700255/1834061/Landsat+ETM%2B%20Data+Format+Control+Book/4bfb7121-e97d-46ca-8d3f-02f1c6bf309c;jsessionid=F97499AA37A1E9AF0EED42900CF66760?version=1.1
+
+    SeeAlso:
+        watch.gis.geotiff.parse_landsat_product_id
 
     TODO:
         - [ ] parametarize scene name / identifier
         - [ ] bundle bands in a single file (gdal VRT?)
-        - [ ] separate data and metadata files in return structure?
+        - [X] separate data and metadata files in return structure?
     """
-    if scene_id is not None:
+    if product_id is not None:
         raise NotImplementedError('Must use the default scene')
 
     scene_name = 'LC08_L1TP_037029_20130602_20170310_01_T1'
@@ -33,8 +61,8 @@ def grab_landsat_item(scene_id=None):
 
     if False:
         # overkill?
-        from watch.gis.geotiff import parse_landsat_scene_name
-        ls_meta = parse_landsat_scene_name(scene_name)
+        from watch.gis.geotiff import parse_landsat_product_id
+        ls_meta = parse_landsat_product_id(scene_name)
         sat_code = ls_meta['sat_code']
     else:
         # Equivalent to above, but more fragile and less dependencies
@@ -49,14 +77,47 @@ def grab_landsat_item(scene_id=None):
 
     # Each satellite has a different set of files that it will produce
     sat_code_to_suffixes = {
-        '01': ['B1.TIF', 'B2.TIF', 'B3.TIF', 'B4.TIF', 'B5.TIF', 'B6.TIF', 'B7.TIF', 'MTL.txt'],
-        '02': ['B1.TIF', 'B2.TIF', 'B3.TIF', 'B4.TIF', 'B5.TIF', 'B6.TIF', 'B7.TIF', 'MTL.txt'],
-        '03': ['B1.TIF', 'B2.TIF', 'B3.TIF', 'B4.TIF', 'B5.TIF', 'B6.TIF', 'B7.TIF', 'MTL.txt'],
-        '04': ['B1.TIF', 'B2.TIF', 'B3.TIF', 'B4.TIF', 'B5.TIF', 'B6.TIF', 'B7.TIF', 'MTL.txt'],
-        '05': ['B1.TIF', 'B2.TIF', 'B3.TIF', 'B4.TIF', 'B5.TIF', 'B6.TIF', 'B7.TIF', 'BQA.TIF', 'MTL.txt'],
-        '07': ['B1.TIF', 'B2.TIF', 'B3.TIF', 'B4.TIF', 'B5.TIF', 'B6_VCID_1.TIF', 'B6_VCID_2.TIF', 'B7.TIF', 'B8.TIF', 'BQA.TIF', 'MTL.txt'],
-        '08': ['B1.TIF', 'B2.TIF', 'B3.TIF', 'B4.TIF', 'B5.TIF', 'B6.TIF', 'B7.TIF', 'B8.TIF', 'B9.TIF', 'B10.TIF', 'B11.TIF', 'BQA.TIF', 'MTL.txt']
+        '01': {
+            'bands': ['B1.TIF', 'B2.TIF', 'B3.TIF', 'B4.TIF', 'B5.TIF', 'B6.TIF', 'B7.TIF'],
+            'meta': {'mtl': 'MTL.txt'}
+        },
+
+        '02': {
+            'bands': ['B1.TIF', 'B2.TIF', 'B3.TIF', 'B4.TIF', 'B5.TIF', 'B6.TIF', 'B7.TIF'],
+            'meta': {'mtl': 'MTL.txt'}
+        },
+
+        '03': {
+            'bands': ['B1.TIF', 'B2.TIF', 'B3.TIF', 'B4.TIF', 'B5.TIF', 'B6.TIF', 'B7.TIF'],
+            'meta': {'mtl': 'MTL.txt', 'bqa': 'BQA.TIF'}
+        },
+
+        '04': {
+            'bands': ['B1.TIF', 'B2.TIF', 'B3.TIF', 'B4.TIF', 'B5.TIF', 'B6.TIF', 'B7.TIF'],
+            'meta': {'mtl': 'MTL.txt', 'bqa': 'BQA.TIF'}
+        },
+
+        '05': {
+            'bands': ['B1.TIF', 'B2.TIF', 'B3.TIF', 'B4.TIF', 'B5.TIF', 'B6.TIF', 'B7.TIF'],
+            'meta': {'mtl': 'MTL.txt', 'bqa': 'BQA.TIF'}
+        },
+
+        '07': {
+            'bands': ['B1.TIF', 'B2.TIF', 'B3.TIF', 'B4.TIF', 'B5.TIF', 'B6_VCID_1.TIF', 'B6_VCID_2.TIF'],
+            'meta': {'mtl': 'MTL.txt', 'bqa': 'BQA.TIF'}
+        },
+
+        '08': {
+            'bands': ['B1.TIF', 'B2.TIF', 'B3.TIF', 'B4.TIF', 'B5.TIF', 'B6.TIF',
+                      'B7.TIF', 'B8.TIF', 'B9.TIF', 'B10.TIF', 'B11.TIF'],
+            'meta': {'mtl': 'MTL.txt', 'bqa': 'BQA.TIF'},
+        }
     }
+
+    # B6_VCID_1 = Band 6 Visual Channel Identifier (VCID) 1
+    # B6_VCID_2 = Band 6 VCID
+    # MTL = metadata
+    # BQA: quality assurance pseudo-band
 
     item_suffixes = sat_code_to_suffixes[sat_code]
 
@@ -66,12 +127,24 @@ def grab_landsat_item(scene_id=None):
     # Cache the scene using the same path used by google cloud storage
     scene_dpath = ub.ensuredir((dset_dpath, scene_path))
 
-    item_fpaths = []
-    for suffix in item_suffixes:
+    product = {
+        'bands': [],
+        'meta': {},
+    }
+    # Download band product-items
+    for suffix in item_suffixes['bands']:
         fname = '{}_{}'.format(scene_name, suffix)
         uri_suffix = join(scene_path, fname)
         item_uri = join(uri_prefix, uri_suffix)
         fpath = ub.grabdata(item_uri, dpath=scene_dpath)
-        item_fpaths.append(fpath)
+        product['bands'].append(fpath)
 
-    return item_fpaths
+    # Download meta product-items
+    for key, suffix in item_suffixes['meta'].items():
+        fname = '{}_{}'.format(scene_name, suffix)
+        uri_suffix = join(scene_path, fname)
+        item_uri = join(uri_prefix, uri_suffix)
+        fpath = ub.grabdata(item_uri, dpath=scene_dpath)
+        product['meta'][key] = fpath
+
+    return product